challenge-1/agents/v2/requirements.txt
#requests
azure-identity
azure-cosmos
agent-framework==1.0.0b251211    
azure-search-documents==11.7.0b2
#azure-ai-projects
# azure-ai-agents
# agent-framework-core
# agent-framework-azure-ai
# azure-ai-projects